requirements.txt
click
datasets
hf_transfer
moviepy
pillow
torch
torchvision
wandb
